项目/钉钉推送/requirements.txt
python-jenkins
requests